plugin: Deal with cluster runtime problems
Before, when you ran the runtime check for a cluster, we'd output all
sorts of information, such as the overall status for the cluster and if
there were any configuration issues, which was great.. however, we didn't
bother to actually use the data, instead opting to just setting the result
to OK every time and ignoring any eventual errors.

This was a bad idea.

It has now been remedied, and cluster checks all over the world will start
alarming, to the horror of their administrators.

Ah.. all in a day's work..

This fixes #6901.

Signed-off-by: Mattias Bergsten <mattias.bergsten@op5.com>