plugin: Bump version number
It's been many years with changes to the plugin but no change to the
version number. This makes it almost impossible to track down which
particular version of the plugin a user is using, which makes
troubleshooting, uh, troublesome.

Hopefully, we can start actually using the version number again, to make
life easier for everyone, mmmkay?

Signed-off-by: Mattias Bergsten <mattias.bergsten@op5.com>