test: introduce test script loading
A script is a sequence of responses as returned by the VMWare web
service for a particular combination of command/subcommands. This patch
introduces functionality for loading the script from a file where each
response is separated by ! (exclamation mark).

This allows us to test commands locally, on well-defined values, without a lot of
manual stubbing and mocking.

This could also be useful for reproducing error conditions, such as
partial responses, and the like.

Signed-off-by: Anton Lofgren <alofgren@op5.com>