generate_test: record output to test script
This allows us to do write tests which in turn allows us to do more aggressive
refactoring, since any functional change to the plugin will cause it to
behave differently from the recorded run.

This is NOT (seriously, NOT (really)) intended as supporting unit tests as such,
since we don't verify that the output is even close to being correct.
This simply verifies that behaviour hasn't changed between two sets of
changes.

However, once we can verify _that_, we can refactor mercilessly and make
this plugin maintainable - which should include some _proper_ unit tests
as well! Hooray!

Signed-off-by: Anton Lofgren <alofgren@op5.com>