build: Add missing dependencies
The custom post test was made against wrong SDK version.
So when changing to latest SDK (5.5) we have missing deps.
This is now fixed and the complete deps list is found inside
the vmware-install.pl script in the array:
@modules
So now all dependencies according to the installation script
are added as requirements in the spec file.

This fixes issue: #7922

Signed-off-by: Martin Kamijo <mk@op5.com>