plugin: Don't specify a default for interval/maxsamples
This has been recorded to avoid the issue of VMware's returning
negative return values for performance queries. I have not been able to
reproduce the issue using this approach on a system where it regularly
occurred otherwise.

Why this is is anyone's guess*, but I believe the best we can do here
is avoidance until VMware properly fixes the bug.

Thus, as much as it hurts me to say this (because it's not _really_
true):

This "fixes" #4532.

However, if someone experiences the issue with this patch applied,
please reopen the bug and provide as much information as possible.

* mine is: "Using the default values (set server-side, instead of our
own) for the parameters avoids some complexity in the query handling
server-side which triggers the bug". Though this is of course impossible
to verify.

Signed-off-by: Anton Lofgren <alofgren@op5.com>