Retrieve real-time performance values
I misread the documentation - I thought the default value was supposed
to be "real-time" (a sampling period of 20 seconds). Turns out, it's
not. The default sampling period seems to be more like, a day - which is
pretty useless for checks on, for example, CPU load.

Hopefully, this doesn't re-trigger the negative-values bug (#4532).

Signed-off-by: Anton Lofgren <alofgren@op5.com>