Remove incorrect uom and add fix for HP perfdata
The allowed uom in perfdata is quite clear in the monitoring-plugin
project, since we would like to follow those guidelines, remove the uom
that does not follow it.

Some HP machines report temperature as ' --- Normal' which is
included in the label name. If the state changes the label in the
perfdata changes thus causing problems for some graphing tools. Fix this
by removing that part from the label in the perfdata.

Signed-off-by: Robin Hagman <robin.hagman@op5.com>