Revert "Session cookies will be used for loading and reusing sessions and not only for faster authentication."
The reverted commit introduced a bug which is presumably fixed in a pull
request at https://github.com/op5/check_vmware_api/pull/6. But we need
to test this part of the code in order to be able to merge it to master.
Reverting this commit for now so that we can test it properly before
changing this part of the code.

This fixes issue MON-9527.

This reverts commit 596940ede042e853c7b65008b01189f9e52607b0.